WIP - Correct tap area with side effects.
I increased the size of the description views to make the whole space tappable.
I dynamically add bottom margin when both description views are missing.

However, The description views are contained within the larger StackView so,
if both are present, they have spacing in between them like the other elements
of the stack view.