Closes #365: Enable firebase persistence.
I tested this by modifying the rating of one place in the firebase and it
updates as expected when opening and closing the app multiple times. A side
effect of this change is how the data is updated - only the place details
screen is updated with the data & the place carousel is not: I filed #475.